Check if menu.1st file exists before moving

+++ misc/install-digitalocean.sh
@@ -1,1 +1,3 @@
-sudo mv /boot/grub/menu.lst /
+if [ -f /boot/grub/menu.lst ]; then
+    sudo mv /boot/grub/menu.lst /
+fi
